Updated tests to work properly.
