+++ tests/test_adaptor.py
@@ -9,8 +9,7 @@ import pytest
 from servicex import ResultDestination
 from servicex.models import ResultFormat, Status, TransformRequest, TransformStatus
 
-from servicex_local import SXLocalAdaptor
-from servicex_local.adaptor import MinioLocalAdaptor
+from servicex_local.adaptor import MinioLocalAdaptor, SXLocalAdaptor
 
 
 def test_adaptor_url():
